docs: Update Sample Netflix Application (#28)
* docs: Update Sample Netflix Application

In this commit we make two changes to the sample Netflix Application. 
1) We change the link to the collab so it opens one for Spanner not CloudSQl (previosly it opened a link to cloud-sql's project)
2) We amend the pip install so that we pip install the Spanner package.

* docs: Update Netflix Sample Application

In this commit we amend the pip install for Spanner's Langchain integration from "langchain_google_spanner" to "langchain-google-spanner"

* docs: Update Netflix Sample Application

In this commit we update the sample Netflix Application notebook. We link to the proper colab url.

* docs: Update Sample Netflix Application

We change to the proper langchain google vertex ai package.

Co-authored-by: Averi Kitsch <akitsch@google.com>

---------

Co-authored-by: Averi Kitsch <akitsch@google.com>