fix: Use Spanner's UPSERT over INSERT (#90)
Switch to UPSERT. This provides for a nice experience for DELETE-then-INSERT use cases in Spanner when the primary key stays the same. Otherwise the DELETE is not pushed before INSERT and the client gets a 409 Row [abc] already exists. The way LangChain integration is designed, there is no easy way to wrap them all in a TX with additional args for such visibility. 

Co-authored-by: goodvibes <goodvibes@google.com>