fix: make ANN algorithm updates based off usage + testing
This change is carved out of PR #138, due to the fact that
at present, Google Cloud Spanner vector index creation and
deletion takes a very long time and is non-deterministic
hence unreliable for integration tests so these updates
which have unit tests are exclusive of the integration tests
and allow for usage directly as Google engineering figures
out the backend issues.

Updates #94.