Include json schema in the graph schema representation
Allow graph schema to include the json fields. This is useful for QA
chain to handle queries that refers to json properties.

For example, for node:
        Company(details = Json('market_cap', ...))
graph schema will include 'market_cap' as a json_fields of `details` in
the schema representation, so that when a user ask for
`get market capitalization of company`, QA chain can understand which
subfield to refer to.

Note:
- json property schema is done via by inspecting the first non-null property.
- other changes: refactor the tests, improve the logging